+++ CHANGELOG.md
@@ -9,4 +9,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 
-[unreleased]: https://github.com/aryelgois/dotfiler/compare/initial-commit...develop
+## [0.1.0] - 2019-10-26
+
+### Added
+
+- Implementation in shell script
+  - `dotfiler init`
+  - `dotfiler mount`
+  - `dotfiler umount`
+  - `dotfiler add`
+  - `dotfiler rm`
+- [Makefile]
+
+
+[unreleased]: https://github.com/aryelgois/dotfiler/compare/v0.1.0...develop
+[0.1.0]: https://github.com/aryelgois/dotfiler/compare/initial-commit...v0.1.0
+
+[makefile]: Makefile